Refactor home page UI and added image, applied minor design updates
