+++ src/pages/NapkinReview.jsx
@@ -35,7 +35,7 @@ const shimmer = keyframes`
 const Container = styled.div`
   max-width: 1200px;
   margin: 0 auto;
-  padding: 0 20px;
+ padding: 50px 25px;
 `;
 
 const PageContainer = styled.div`
